pyproject.toml
[project]
name = "ppt-parser"
version = "0.1.0"
description = "PPTX to SQLite parser and web viewer"
requires-python = ">=3.10"
dependencies = [
    "python-pptx>=1.0.0",
    "lxml>=4.9.0",
    "Pillow>=10.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
]

[project.scripts]
ppt-parse = "parser.cli:main"
